refactor: Reorganize
