+++ scripts/banking/deleteDemoData.js
