+++ scripts/banking/deleteDuplicateBankAccountsWithNoOperations.spec.js
