+++ scripts/banking/deleteOrphanOperations.spec.js
@@ -1,4 +1,4 @@
-const api = require('./api')
+const api = require('../api')
 const { run } = require('./deleteOrphanOperations')
 
 const makeDoc = (id, doc) => ({ id, doc: { _id: id, ...doc } })
